added generic parser and refactored miseq class

+++ mytd_parser/parse_server_copy.py
@@ -13,16 +13,15 @@ from .logger_settings import api_logger
 from pathlib import *
 import datetime
 from datetime import datetime
-from mytd_parser.parse_seq_run import MiSeqParser, get_creation_dt
+from mytd_parser.parse_seq_run import get_creation_dt, check_rta_complete, get_run_id_xml
 
 
-class ServerParse(MiSeqParser):
+class ServerParse:
     def __init__(self, download_dir=settings.SERVER_DOWNLOAD_DIR,
                  output_dir=settings.SERVER_OUTPUT_DIR,
                  data_directory=settings.SERVER_DATA_DIRECTORY,
-                 upload_bioinfo_results_dir=settings.SERVER_UPLOAD_BR_DIR,
-                 log_file_dir=settings.LOG_FILE_DIR):
-        super().__init__(download_dir, output_dir, data_directory)
+                 log_file_dir=settings.LOG_FILE_DIR,
+                 upload_bioinfo_results_dir=settings.SERVER_UPLOAD_BR_DIR):
         self.download_dir = download_dir
         self.output_dir = output_dir
         self.upload_bioinfo_results_dir = upload_bioinfo_results_dir
@@ -53,6 +52,7 @@ class ServerParse(MiSeqParser):
          get dirs and put in pandas df
         """
         try:
+            server_parse_dates = []
             run_ids = []
             run_dirs = []
             fastq_dirs = []
@@ -65,6 +65,7 @@ class ServerParse(MiSeqParser):
             analysis_completes = []
             upload_completes = []
             # make list of all project folders; e.g., maine-edna
+            server_parse_date = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
             project_dirs = glob.glob(os.path.join(self.download_dir, '*/'))
             for project_dir in project_dirs:
                 dir_length = len(os.listdir(project_dir))
@@ -73,8 +74,8 @@ class ServerParse(MiSeqParser):
                     continue
                 project = Path(project_dir).name
                 # grab sequencing folder directory with most recent date for each project
-                #run_dir = max(glob.glob(os.path.join(project_dir, '*/')), key=os.path.getmtime)
-                #run_dir = run_dir.replace('\\', '/')
+                # run_dir = max(glob.glob(os.path.join(project_dir, '*/')), key=os.path.getmtime)
+                # run_dir = run_dir.replace('\\', '/')
                 # make list of all sequencing run folders in project directories
                 run_dir_list = glob.glob(os.path.join(project_dir, '*/'))
                 run_dir_list = [dir_path.replace('\\', '/') for dir_path in run_dir_list]
@@ -84,9 +85,9 @@ class ServerParse(MiSeqParser):
                     num_run_dir = len(run_dir_list)
                     # if rta_complete exists, sets variable to True to be filtered on
                     # to only process rta_complete directories
-                    rta_complete = self.check_rta_complete(run_dir)
+                    rta_complete = check_rta_complete(run_dir)
                     # grab run_id from CompleteJobInfo.xml
-                    run_id = self.get_run_id_xml(run_dir)
+                    run_id = get_run_id_xml(run_dir)
                     run_dir_create_date = datetime.fromtimestamp(get_creation_dt(run_dir)).strftime('%Y-%m-%d %H:%M:%S')
                     if not rta_complete:
                         # if rta_complete is false, then the run is not complete
@@ -96,6 +97,7 @@ class ServerParse(MiSeqParser):
                         upload_complete = False
 
                         # append to lists
+                        server_parse_dates.append(server_parse_date)
                         projects.append(project)
                         run_ids.append(run_id)
                         run_dirs.append(run_dir)
@@ -122,6 +124,7 @@ class ServerParse(MiSeqParser):
                             upload_complete = False
 
                             # append to lists
+                            server_parse_dates.append(server_parse_date)
                             projects.append(project)
                             run_ids.append(run_id)
                             run_dirs.append(run_dir)
@@ -146,6 +149,7 @@ class ServerParse(MiSeqParser):
                                     fastq_dir_create_date = datetime.fromtimestamp(get_creation_dt(fastq_dir)).strftime('%Y-%m-%d %H:%M:%S')
                                     analysis_complete = True
                                     # append to lists
+                                    server_parse_dates.append(server_parse_date)
                                     projects.append(project)
                                     run_ids.append(run_id)
                                     run_dirs.append(run_dir)
@@ -164,6 +168,7 @@ class ServerParse(MiSeqParser):
                                     fastq_dir_create_date = datetime.fromtimestamp(get_creation_dt(fastq_dir)).strftime('%Y-%m-%d %H:%M:%S')
                                     analysis_complete = True
                                     # append to lists
+                                    server_parse_dates.append(server_parse_date)
                                     projects.append(project)
                                     run_ids.append(run_id)
                                     run_dirs.append(run_dir)
@@ -176,22 +181,26 @@ class ServerParse(MiSeqParser):
                                     analysis_completes.append(analysis_complete)
                                     upload_completes.append(upload_complete)
 
-            dirs_df = pd.DataFrame(list(zip(projects, run_ids, run_dirs, run_dirs_create_dates, num_run_dirs,
-                                            fastq_dirs, fastq_dirs_create_dates, num_fastq_dirs, rta_completes,
-                                            analysis_completes, upload_completes)),
-                                   columns=['project', 'run_id', 'run_dir','run_dir_create_date', 'num_run_dir',
-                                            'fastq_dir', 'fastq_dir_create_date', 'num_fastq_dir', 'rta_complete',
-                                            'analysis_complete', 'upload_complete'])
+            dirs_df = pd.DataFrame(list(zip(server_parse_dates, projects, run_ids, run_dirs, run_dirs_create_dates,
+                                            num_run_dirs, fastq_dirs, fastq_dirs_create_dates, num_fastq_dirs,
+                                            rta_completes, analysis_completes, upload_completes)),
+                                   columns=['server_parse_date', 'project', 'run_id', 'run_dir', 'run_dir_create_date',
+                                            'num_run_dir', 'fastq_dir', 'fastq_dir_create_date', 'num_fastq_dir',
+                                            'rta_complete', 'analysis_complete', 'upload_complete'])
             # output dirs to csv
             if export_csv:
-                output_csv_filename = datetime.now().strftime(self.log_file_dir + 'server_dirlist_%Y%m%d_%H%M%S.csv')
-                dirs_df.to_csv(output_csv_filename, encoding='utf-8', index=False)
+                output_csv_filename = self.log_file_dir + 'server_dirlist.csv'
+                if os.path.exists(output_csv_filename):
+                    dirs_df.to_csv(output_csv_filename, encoding='utf-8', index=False, header=False, mode='a')
+                else:
+                    dirs_df.to_csv(output_csv_filename, encoding='utf-8', index=False)
             if complete_upload:
                 # subset by directories that have RTAComplete.txt; we do not want to process incomplete sequencing runs
-                dirs_df_upload_complete = dirs_df[(dirs_df['rta_complete'] == True) & (dirs_df['upload_complete'] == True)]
-                return(dirs_df_upload_complete)
+                dirs_df_upload_complete = dirs_df[(dirs_df['rta_complete'] == True) &
+                                                  (dirs_df['upload_complete'] == True)]
+                return dirs_df_upload_complete
             else:
-                return(dirs_df)
+                return dirs_df
         except Exception as err:
             raise RuntimeError("** Error: get_dirs Failed (" + str(err) + ")")
 
@@ -202,7 +211,7 @@ class ServerParse(MiSeqParser):
         try:
             api_logger.info('[START] create_bioinformatics_results_dir')
             upload_bioinfo_results_dir = self.upload_bioinfo_results_dir
-            output_hpc_run_dir = upload_bioinfo_results_dir+project+'/'+run_id+'/Bioinformatics_Results/'
+            output_hpc_run_dir = upload_bioinfo_results_dir + project + '/' + run_id + '/Bioinformatics_Results/'
             api_logger.info('Start create dir: ' + project + ', ' + run_id + ': [' + output_hpc_run_dir + ']')
             if not os.path.exists(output_hpc_run_dir):
                 os.makedirs(output_hpc_run_dir)
@@ -223,23 +232,24 @@ class ServerParse(MiSeqParser):
                 run_id = row['run_id']
                 fastq_dir = row['fastq_dir']
                 self.create_bioinformatics_results_dir(project, run_id)
-                output_fastq_dir = output_dir+run_id+'/'
+                output_fastq_dir = output_dir + run_id + '/'
                 if not os.path.exists(output_fastq_dir):
                     os.makedirs(output_fastq_dir)
                 fastq_files = glob.glob(os.path.join(fastq_dir, '**/*.fastq.gz'), recursive=True)
                 num_fastq_files = len(fastq_files)
-                api_logger.info('Start move: '+project+', '+run_id+', '+str(num_fastq_files)+' files from: [' + fastq_dir + '], to: [' + output_fastq_dir + ']')
-                fastq_counter=0
+                api_logger.info('Start move: ' + project + ', ' + run_id + ', ' + str(num_fastq_files) +
+                                ' files from: [' + fastq_dir + '], to: [' + output_fastq_dir + ']')
+                fastq_counter = 0
                 for fastq_file in fastq_files:
                     fastq_filename = os.path.basename(fastq_file)
-                    output_fastq_filename = output_fastq_dir+fastq_filename
+                    output_fastq_filename = output_fastq_dir + fastq_filename
                     if not os.path.exists(output_fastq_filename):
                         # only want to copy/move if file doesn't already exist
                         copy2(fastq_file, output_fastq_dir)
                         api_logger.info(str(fastq_filename) + ' moved to [' + str(output_fastq_dir) + ']')
                         fastq_counter += 1
                     else:
-                        api_logger.info(str(fastq_filename)+' already in ['+str(output_fastq_dir)+']')
+                        api_logger.info(str(fastq_filename) + ' already in [' + str(output_fastq_dir)+']')
                 api_logger.info('End: moved ' + str(fastq_counter) + ' files')
             api_logger.info('[END] move_fastq_files')
         except Exception as err:
